+++ code/test/sixsq/nuvla/server/resources/nuvlabox_1_lifecycle_test.clj
@@ -75,7 +75,8 @@
                      :form-factor      "Nuvlabox"
                      :vm-cidr          "10.0.0.0/24"
                      :lan-cidr         "10.0.1.0/24"
-                     :ssh-keys         ["credential/aaa-bbb-ccc"]})
+                     :ssh-keys         ["credential/aaa-bbb-ccc"]
+                     :capabilities     ["RANDOM"]})
 
 
 (deftest check-metadata
@@ -381,7 +382,8 @@
                                                 :tags                tags
                                                 :minio-access-key    "access"
                                                 :minio-secret-key    "secret"
-                                                :minio-endpoint      "https://minio.example.com"}))
+                                                :minio-endpoint      "https://minio.example.com"
+                                                :capabilities        ["NUVLA_JOB_PULL"]}))
                 (ltu/body->edn)
                 (ltu/is-status 200))
 
@@ -446,19 +448,19 @@
             (-> session
                 (request commission
                          :request-method :post
-                         :body (json/write-str {:swarm-token-worker  "abc"
-                                                :swarm-token-manager "def"
-                                                :swarm-client-key    "key"
-                                                :swarm-client-cert   "cert"
-                                                :swarm-client-ca     "ca"
-                                                :swarm-endpoint      "https://swarm.example.com"
-                                                :minio-access-key    "access"
-                                                :minio-secret-key    "secret"
-                                                :minio-endpoint      "https://minio.example.com"
-                                                :kubernetes-client-key    "key"
-                                                :kubernetes-client-cert   "cert"
-                                                :kubernetes-client-ca     "ca"
-                                                :kubernetes-endpoint      "https://k8s.example.com"}))
+                         :body (json/write-str {:swarm-token-worker     "abc"
+                                                :swarm-token-manager    "def"
+                                                :swarm-client-key       "key"
+                                                :swarm-client-cert      "cert"
+                                                :swarm-client-ca        "ca"
+                                                :swarm-endpoint         "https://swarm.example.com"
+                                                :minio-access-key       "access"
+                                                :minio-secret-key       "secret"
+                                                :minio-endpoint         "https://minio.example.com"
+                                                :kubernetes-client-key  "key"
+                                                :kubernetes-client-cert "cert"
+                                                :kubernetes-client-ca   "ca"
+                                                :kubernetes-endpoint    "https://k8s.example.com"}))
                 (ltu/body->edn)
                 (ltu/is-status 200))
 
@@ -488,7 +490,8 @@
                                                 :swarm-endpoint      "https://swarm.example.com"
                                                 :minio-access-key    "access"
                                                 :minio-secret-key    "secret"
-                                                :minio-endpoint      "https://minio.example.com"}))
+                                                :minio-endpoint      "https://minio.example.com"
+                                                :capabilities        []}))
                 (ltu/body->edn)
                 (ltu/is-status 200))
 
@@ -686,19 +689,19 @@
           (-> session-owner
               (request commission
                        :request-method :post
-                       :body (json/write-str {:swarm-token-worker  "abc"
-                                              :swarm-token-manager "def"
-                                              :swarm-client-key    "key"
-                                              :swarm-client-cert   "cert"
-                                              :swarm-client-ca     "ca"
-                                              :swarm-endpoint      "https://swarm.example.com"
-                                              :minio-access-key    "access"
-                                              :minio-secret-key    "secret"
-                                              :minio-endpoint      "https://minio.example.com"
-                                              :kubernetes-client-key    "key"
-                                              :kubernetes-client-cert   "cert"
-                                              :kubernetes-client-ca     "ca"
-                                              :kubernetes-endpoint      "https://k8s.example.com"}))
+                       :body (json/write-str {:swarm-token-worker     "abc"
+                                              :swarm-token-manager    "def"
+                                              :swarm-client-key       "key"
+                                              :swarm-client-cert      "cert"
+                                              :swarm-client-ca        "ca"
+                                              :swarm-endpoint         "https://swarm.example.com"
+                                              :minio-access-key       "access"
+                                              :minio-secret-key       "secret"
+                                              :minio-endpoint         "https://minio.example.com"
+                                              :kubernetes-client-key  "key"
+                                              :kubernetes-client-cert "cert"
+                                              :kubernetes-client-ca   "ca"
+                                              :kubernetes-endpoint    "https://k8s.example.com"}))
               (ltu/body->edn)
               (ltu/is-status 200))
 
@@ -1166,6 +1169,159 @@
             stripe/delete-customer))
 
       )))
+
+
+(deftest execution-mode-action-lifecycle
+  (binding [config-nuvla/*stripe-api-key* nil]
+    (let [session       (-> (ltu/ring-app)
+                            session
+                            (content-type "application/json"))
+
+          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
+          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
+
+          nuvlabox-id   (-> session-owner
+                            (request base-uri
+                                     :request-method :post
+                                     :body (json/write-str valid-nuvlabox))
+                            (ltu/body->edn)
+                            (ltu/is-status 201)
+                            (ltu/location))
+          nuvlabox-url  (str p/service-context nuvlabox-id)]
+
+      (-> session-admin
+          (request nuvlabox-url
+                   :request-method :put
+                   :body (json/write-str {:state "COMMISSIONED"}))
+          (ltu/body->edn)
+          (ltu/is-status 200))
+
+      ;; push
+      (let [reboot-url (-> session-owner
+                           (request nuvlabox-url)
+                           (ltu/body->edn)
+                           (ltu/is-status 200)
+                           (ltu/get-op-url :reboot))]
+        (let [job-url    (-> session-owner
+                             (request reboot-url :request-method :post)
+                             (ltu/body->edn)
+                             (ltu/is-status 202)
+                             (ltu/location-url))]
+         (-> session-admin
+             (request job-url)
+             (ltu/body->edn)
+             (ltu/is-status 200)
+             (ltu/is-key-value :execution-mode "push"))
+
+         (-> session-admin
+             (request job-url
+                      :request-method :put
+                      :body (json/write-str {:execution-mode "pull"}))
+             (ltu/body->edn)
+             (ltu/is-status 200)
+             (ltu/is-key-value :execution-mode "pull")
+             (ltu/is-key-value :acl {:edit-data ["user/alpha"],
+                                     :owners    ["group/nuvla-admin"],
+                                     :view-acl  ["user/alpha"],
+                                     :delete    ["user/alpha"],
+                                     :view-meta [nuvlabox-id "user/alpha"],
+                                     :edit-acl  ["user/alpha"],
+                                     :view-data [nuvlabox-id "user/alpha"],
+                                     :manage    [nuvlabox-id "user/alpha"],
+                                     :edit-meta ["user/alpha"]}))))
+
+      )))
+
+
+(deftest create-activate-commission-get-context-lifecycle
+  (binding [config-nuvla/*stripe-api-key* nil]
+    (let [session       (-> (ltu/ring-app)
+                            session
+                            (content-type "application/json"))
+          session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
+
+          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
+          session-anon  (header session authn-info-header "unknown group/nuvla-anon")]
+
+      (let [nuvlabox-id  (-> session-owner
+                             (request base-uri
+                                      :request-method :post
+                                      :body (json/write-str valid-nuvlabox))
+                             (ltu/body->edn)
+                             (ltu/is-status 201)
+                             (ltu/location))
+
+            nuvlabox-url (str p/service-context nuvlabox-id)
+
+            activate-url (-> session-owner
+                             (request nuvlabox-url)
+                             (ltu/body->edn)
+                             (ltu/is-status 200)
+                             (ltu/get-op-url :activate))]
+
+        ;; activate nuvlabox
+        (-> session-anon
+            (request activate-url
+                     :request-method :post)
+            (ltu/body->edn)
+            (ltu/is-status 200))
+
+        (let [session-nuvlabox (header session authn-info-header
+                                       (str nuvlabox-id
+                                            " group/nuvla-nuvlabox group/nuvla-anon"))
+              commission       (-> session-owner
+                                   (request nuvlabox-url)
+                                   (ltu/body->edn)
+                                   (ltu/is-status 200)
+                                   (ltu/get-op-url :commission))]
+
+          (-> session-nuvlabox
+              (request commission
+                       :request-method :post)
+              (ltu/body->edn)
+              (ltu/is-status 200))
+
+          (let [add-ssh-key     (-> session-owner
+                                    (request nuvlabox-url)
+                                    (ltu/body->edn)
+                                    (ltu/is-status 200)
+                                    (ltu/get-op-url :add-ssh-key))
+                job-url         (-> session-owner
+                                    (request add-ssh-key
+                                             :request-method :post)
+                                    (ltu/body->edn)
+                                    (ltu/is-status 202)
+                                    (ltu/location-url))
+
+                get-context-url (-> session-nuvlabox
+                                    (request job-url)
+                                    (ltu/body->edn)
+                                    (ltu/is-status 200)
+                                    (ltu/is-operation-present :get-context)
+                                    (ltu/get-op-url :get-context))
+
+                get-context-body (-> session-nuvlabox
+                                     (request get-context-url)
+                                     (ltu/body->edn)
+                                     (ltu/is-status 200)
+                                     (ltu/body))]
+
+            (is (and (map? get-context-body)
+                     (str/starts-with? (namespace (ffirst get-context-body)) "credential")
+                     (str/starts-with? (:public-key (second (first get-context-body)))
+                                       "ssh-rsa AAAA")))
+
+            (-> session-admin
+                (request get-context-url)
+                (ltu/body->edn)
+                (ltu/is-status 200))
+
+            (-> session-owner
+                (request get-context-url)
+                (ltu/body->edn)
+                (ltu/is-status 403)))
+
+          )))))
 
 
 (deftest bad-methods